Error fixes.
Added a cancel button test to the antenna progress dialog, and removed the stop button from the antenna menu.
Fixed the S-Parameters function as the VNA was giving query interrupted errors.
Adjusted the app, removed unnecessary code, and fixed minor bugs with the homing and scanning of the linear slider.